refactor(core): split subrequest.rs into submodules (#963)
* refactor(core): split subrequest.rs into submodules

Break the 1491-line monolith into focused submodules:

  subrequest/
    mod.rs        — module doc + re-exports
    types.rs      — SubRequest, SubResponse, FrameworkHeaders, SubRequestError
    client.rs     — SubRequestConnector, SubRequestClient, execute_inner, CircuitGuard
    internals.rs  — header sanitization + timeout helpers
    tests.rs      — all unit tests

Public API is unchanged; all existing imports continue to work
through the re-exports in mod.rs.


* docs: update stale subrequest path in proposal doc

The file moved from core/src/subrequest.rs to
core/src/subrequest/client.rs in the module split.


* fix(core): restore design-rationale comment for Unix socket circuit breaker exclusion


---------
